Added Nav Bar
I have added the nav bar for this store. I copied this navbar from one of my previous tailwind projects to keep things consistent. However, I did change the icon in the top left. I also removed the image, because it doesn't make sense for a user to have an image on a store website.