Added/fixed Images
I have added 2 new images to the new sale page. I have also updated one of the cards to show some disinfecting wipes that are part of the essentials sale. Lastly, there was an issue with the logo that it wasn't taking you to the main homepage, so I fixed that.