Fixed Spacing
There was an issue with centering and layout on the cards. I have since fixed this. The image and text now each take up half the card. Also, the text side is vertically centered. This fix will need to be added to the product on the main page as well.