Added Data
I have added both a data.json file and an images folder into a new folder titled 'data'. This is the sample data I will be using. However, this could also work with any other API. You would just need to call it and the data would be handled the same way.